feat(api): proxy new format
