+++ packages/server/lib/controllers/proxy/allProxy.unit.test.ts
@@ -1,11 +1,11 @@
 import { describe, expect, it } from 'vitest';
 
-import { parseHeaders } from './proxy.controller.js';
+import { parseHeaders } from './allProxy.js';
 
 import type { Request } from 'express';
 
-describe('Proxy Controller Construct URL Tests', () => {
-    it('Should parse headers that starts with Nango-Proxy or nango-proxy', () => {
+describe('parseHeaders', () => {
+    it('should parse headers that starts with Nango-Proxy or nango-proxy', () => {
         const req: Pick<Request, 'rawHeaders'> = {
             rawHeaders: ['Nango-Proxy-Test-Header', 'TestValue', 'nango-proxy-another-header', 'AnotherValue', 'Irrelevant-Header', 'IrrelevantValue']
         };
@@ -18,7 +18,7 @@ describe('Proxy Controller Construct URL Tests', () => {
         });
     });
 
-    it('Should return an empty object when there are no Nango-Proxy or nango-proxy headers', () => {
+    it('should return an empty object when there are no Nango-Proxy or nango-proxy headers', () => {
         const req: Pick<Request, 'rawHeaders'> = {
             rawHeaders: ['Irrelevant-Header-One', 'IrrelevantValueOne', 'Irrelevant-Header-Two', 'IrrelevantValueTwo']
         };
@@ -28,7 +28,7 @@ describe('Proxy Controller Construct URL Tests', () => {
         expect(parsedHeaders).toEqual({});
     });
 
-    it('Should handle the case when rawHeaders is not an array or empty', () => {
+    it('should handle the case when rawHeaders is not an array or empty', () => {
         const req = {};
 
         const parsedHeaders = parseHeaders(req as Pick<Request, 'rawHeaders'>);
